Added posix::sys_stream{}

+++ include/rua/log.hpp
@@ -22,9 +22,9 @@ public:
 		_tit(u8_to_w(title)),
 		_ico(icon) {}
 
-	virtual size_t write(bytes_view p) {
+	virtual ptrdiff_t write(bytes_view p) {
 		MessageBoxW(0, u8_to_w(p).c_str(), _tit.c_str(), _ico);
-		return p.size();
+		return static_cast<ptrdiff_t>(p.size());
 	}
 
 private:
